git-server: Don't create a git repo by default
With this change[1] to syncctl the default repo isn't used anymore, so
don't create it by default.

The tests have been tweaked as chart-testing-action doesn't support
providing a custom values file for testing[3] (at least not in a clean
way), so create the git repository on disk like syncctl[4] before
testing pushing.

[1] 96a28d1 ("Support Flux instead of Yggdrasil[1]/Argo CD")[2]
[2] https://github.com/distributed-technologies/syncctl/commit/96a28d16dd08df8296dec3c158c870a53503108d
[3] https://github.com/helm/chart-testing-action/issues/47
[4] https://github.com/distributed-technologies/syncctl/pull/7