artifact-registry: Revert to configuration supported by kind
The local-storage storage class requires the PV to be provisioned
manually (useful for testing), but the values shouldn't be tweaked for
testing, but for production workloads. This partly reverts the change
in[1].

This breaks the test, but they are likely already broken due to[2], and
the chart should likely be rewritten anyways.

[1] 68e15de ("Added test for helm server")
[2] 1058812 ("Remove test folder artifact-registry")