tenant-base: Add support for external secrets
This adds support for external secrets, which is the way we have decided
to handle secrets. The user must provide a name for the external secret
and the key (eg.: path in Vault). The keys in the secret are set as
environment variables in the container.

The user can set secretStore, but this is expected to be handled by a
Kustomization patch.