Replace homebrewed dependency updater with Renovate
We are already using Renovate for some of our other repos and it is
standard tool, so let's also use it here.

For now we just want renovate to update helm dependencies. It can also
update dockerfiles, github actions and helm values files and perhaps we
want to use that functionality in the future.