tenant-base: Ensure the cm name is not longer than 64 chars
When the configmap is digested by Kyverno the cm name is set as a lable value on the generated resource.
This value cannot be more than 64 chars and therefore we make a non random id from the application name and the kafka topic name, that is assured to not be more than 64 chars.